Removed compiler warnings.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@4689 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/gui/common/RsCollectionDialog.cpp
@@ -130,7 +130,7 @@ void RsCollectionDialog::updateSizes()
 	uint64_t total_size = 0 ;
 	uint32_t total_files = 0 ;
 
-	for(int i=0;i<_dlinfos.size();++i)
+	for(size_t i=0;i<_dlinfos.size();++i)
 		if(_cboxes[i]->isChecked())
 		{
 			total_size += _dlinfos[i].size ;
@@ -143,14 +143,14 @@ void RsCollectionDialog::updateSizes()
 void RsCollectionDialog::selectAll() const
 {
 	std::cerr << "Selecting all !" << std::endl;
-	for(int i=0;i<_dlinfos.size();++i)
+	for(size_t i=0;i<_dlinfos.size();++i)
 		_cboxes[i]->setChecked(true) ;
 }
 
 void RsCollectionDialog::deselectAll() const
 {
 	std::cerr << "Deselecting all !" << std::endl;
-	for(int i=0;i<_dlinfos.size();++i)
+	for(size_t i=0;i<_dlinfos.size();++i)
 		_cboxes[i]->setChecked(false) ;
 }
